ch07/iac-security/requirements.txt
click>=8.1.0
pytest>=8.0.0
pytest-cov>=5.0.0
# PyYAML for CloudFormation YAML template parsing
PyYAML>=6.0
